+++ src/cmd/validate/args.rs
@@ -1,0 +1,7 @@
+use clap::Args;
+
+#[derive(Args)]
+pub struct ValidateArgs {
+    /// Version to check
+    pub version: String,
+}